==> 7.Describe_and_create_filtered_object/5.Assess.Duprate.Final.R <==
##### ---> Duplication rates
##### 

## libraries
library(vroom)
library(foreach)
library(tidyverse)
library(magrittr)
library(data.table)


samps <- vroom("dest_v2.samps_25Feb2023.csv")
setDT(samps)
samps = samps[set!="dgn"]

qcs <- system("ls /project/berglandlab/DEST/dest_mapped/*/*/*duplicates_report.txt", intern=T)
#qcs <- system("ls /project/berglandlab/DEST/dest_mapped/DEST_plus/*/*duplicates_report.txt", intern=T)

dup.rat.head <- read.table( pipe( paste("sed -n 7p ", 
                                        paste(qcs[1]) 
                                        )
                                  ))
####
dup.rate <- foreach(i=1:length(qcs), .combine = "rbind" )%do%{
  
  
  dup.rat <- read.table( pipe( paste("sed -n 8p ", 
                                     paste(qcs[i]))
  ))
  
  dup.rat %<>% as.data.frame() %>% mutate(samp = str_split( qcs[i], "/")[[1]][7] )
  message(paste(i, dim(dup.rat)[2], sep = "-") )
  return(data.frame(sampleId = dup.rat$samp,
                    pcrdup = dup.rat$V9))
}

#names(dup.rate)[1:10] = dup.rat.head

save(dup.rate, file = "DuplicateRates.all.Rdata")
